// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  VERIFIED_USER: 'VERIFIED_USER',
  BANK_SAMPAH_ADMIN: 'BANK_SAMPAH_ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const BankSampahStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type BankSampahStatus = (typeof BankSampahStatus)[keyof typeof BankSampahStatus]


export const OlahanStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type OlahanStatus = (typeof OlahanStatus)[keyof typeof OlahanStatus]